IIWA-197: Update common interfaces
