+++ src/backend-api/Helpers/HardwareInterfaces.cs
@@ -2,23 +2,15 @@
 
 namespace backend_api.Helpers
 {
-    // Interface for archiving hardware
-    public interface ISoftDeletable
-    {
-        bool IsDeleted { get; set; }
-    }
-
-    // Interface for assigning hardware
-    public interface IAssignable
-    {
-        int? EmployeeId { get; set; }
-        bool IsAssigned { get; set; }
-    }
 
     // Interface of attributes shared among hardware types.
-    public interface IHardwareBase : IAssignable, ISoftDeletable
+    public interface IHardwareBase : IAssignable, ISoftDeletable, IPurcahseRenewal
     {
-        int GetId();
+        decimal? CostPerYear { get; set; }
+        decimal? FlatCost { get; set; }
         DateTime? PurchaseDate { get; set; }
+        int GetId();
+
     }
+
 }
